pyproject.toml
[project]
name = "sql-auto-optimizer"
version = "0.1.0"
description = "Machine learning-based SQL query optimization using TPC-H benchmark"
authors = [
    { name="Austin Mitchell", email="amitchell80@gatech.edu" }
]
dependencies = [
    "psycopg2-binary>=2.9",
    "pandas>=1.5",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "jupyter>=1.0",
    "xgboost",
    "sentence_transformers",
    "typer[all]>=0.9",  # Optional for CLI scripting
]

[tool.setuptools]
packages = ["src"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
